Segregate discount types
