+++ config/doctrine/ZfrCash.Entity.AbstractDiscount.dcm.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<doctrine-mapping
+        xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
+    http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+
+    <mapped-superclass name="ZfrCash\Entity\AbstractDiscount">
+        <id name="id" type="integer">
+            <generator strategy="AUTO" />
+        </id>
+
+        <one-to-many field="customer" target-entity="ZfrCash\Entity\CustomerInterface" mapped-by="discount" />
+        <one-to-one field="subscription" target-entity="ZfrCash\Entity\Subscription" mapped-by="discount" />
+        <embedded name="coupon" class="ZfrCash\Entity\Coupon" />
+        <field name="startedAt" column="started_at" type="datetime" />
+        <field name="endAt" column="end_at" type="datetime" nullable="true" />
+    </mapped-superclass>
+</doctrine-mapping>
