+++ config/doctrine/ZfrCash.Entity.CustomerDiscount.dcm.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<doctrine-mapping
+        xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
+    http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+
+    <entity name="ZfrCash\Entity\CustomerDiscount" table="cash_customer_discounts">
+        <one-to-one field="customer" target-entity="ZfrCash\Entity\CustomerInterface" mapped-by="discount" />
+    </entity>
+</doctrine-mapping>
