+++ config/doctrine/ZfrCash.Entity.SubscriptionDiscount.dcm.xml
@@ -5,15 +5,8 @@
         xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
     http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-    <entity name="ZfrCash\Entity\Discount" table="cash_discounts">
-        <id name="id" type="integer">
-            <generator strategy="AUTO" />
-        </id>
-
-        <one-to-many field="customer" target-entity="ZfrCash\Entity\CustomerInterface" mapped-by="discount" />
+    <entity name="ZfrCash\Entity\SubscriptionDiscount" table="cash_subscription_discounts">
+        <one-to-many field="customer" target-entity="ZfrCash\Entity\CustomerInterface" />
         <one-to-one field="subscription" target-entity="ZfrCash\Entity\Subscription" mapped-by="discount" />
-        <embedded name="coupon" class="ZfrCash\Entity\Coupon" />
-        <field name="startedAt" column="started_at" type="datetime" />
-        <field name="endAt" column="end_at" type="datetime" nullable="true" />
     </entity>
 </doctrine-mapping>
